# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JCEmpleos API",
        "description": "Documentación oficial del API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Login de usuario",
                "operationId": "44212a9096e4b09358281e9ec8a0701d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@correo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "79854151e609bd20032aa5aa4d5a35c0",
                "responses": {
                    "200": {
                        "description": "Logout exitoso"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener usuario autenticado",
                "operationId": "f606b7972b0f95109c9cd5b591061f41",
                "responses": {
                    "200": {
                        "description": "Usuario autenticado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/change-password": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cambiar contraseña",
                "operationId": "22333fadd20145176e61fc379adc1d7a",
                "responses": {
                    "200": {
                        "description": "Contraseña actualizada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registro de usuario postulante",
                "operationId": "a718f172ff3ac464d723835815f8fb57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellido",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "Password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario creado correctamente"
                    },
                    "422": {
                        "description": "Error de validación o correo duplicado"
                    }
                }
            }
        },
        "/api/perfil": {
            "get": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Obtener perfil del usuario autenticado",
                "operationId": "9125bb1088b7de877a3b12ebdfac0256",
                "responses": {
                    "200": {
                        "description": "Perfil obtenido correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Actualizar perfil del usuario autenticado",
                "operationId": "3f159971c66f848f2435b1ca72b0d993",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pais_id": {
                                        "type": "integer"
                                    },
                                    "departamento_id": {
                                        "type": "integer"
                                    },
                                    "ciudad_id": {
                                        "type": "integer"
                                    },
                                    "sexo_id": {
                                        "type": "integer"
                                    },
                                    "nacionalidad_id": {
                                        "type": "integer"
                                    },
                                    "disponibilidad_vehicular_id": {
                                        "type": "integer"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "example": "1995-05-20"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "foto": {
                                        "type": "string"
                                    },
                                    "acerca_de_mi": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/perfil/cv": {
            "get": {
                "tags": [
                    "Perfil - CV"
                ],
                "summary": "Obtener CV completo del usuario",
                "operationId": "55fd95baaa4c18ce3c135c8c2814c8ed",
                "responses": {
                    "200": {
                        "description": "CV obtenido correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}